index: add the useful thread count to compressed index header lines
when we use '<' to shrink the header (to see more of the subject line)
while viewing the message index, we want to make sure to see the message
count in threads because it's exceedingly useful.  so add it to the
shortened version (it's already in the long version i.e. '>')